+++ samples/scrapping/index.ts
@@ -1,6 +1,6 @@
 import { ReadableApp } from "@scramjet/types";
 import { PassThrough } from "stream";
-import scrap from "./cheerio";
+import scrap from "./scrapper";
 
 
 const app: ReadableApp<string> = async function (_stream: any) {
